move gson files to vendor folder
