+++ sentry/src/main/java/io/sentry/JsonDeserializer.java
@@ -1,9 +1,9 @@
-package io.sentry.json;
+package io.sentry;
 
 import org.jetbrains.annotations.ApiStatus;
 import org.jetbrains.annotations.NotNull;
 
-import io.sentry.json.stream.JsonReader;
+import io.sentry.vendor.gson.stream.JsonReader;
 
 @ApiStatus.Internal
 public interface JsonDeserializer<T> {
